backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pdfplumber==0.11.0
python-pptx==0.6.23
requests==2.31.0
python-dotenv==1.0.1
json-repair==0.23.0
bcrypt==4.2.0
PyJWT==2.9.0
email-validator==2.2.0
